refactor(#5015): keep runtime mismatch seam in provider module
