+++ ARCHITECTURE.md
@@ -731,7 +731,6 @@ src/
 │   │   │   │   ├── intake_turn.rs
 │   │   │   │   ├── latency_spans.rs
 │   │   │   │   ├── provider_isolation.rs
-│   │   │   │   ├── runtime_mismatch.rs
 │   │   │   │   ├── session_strategy_lifecycle_tests.rs
 │   │   │   │   ├── tui_followup.rs
 │   │   │   │   ├── tui_followup_retry.rs
