+++ src/services/discord/router/message_handler.rs
@@ -51,7 +51,6 @@ mod headless_turn;
 mod intake_turn;
 mod latency_spans;
 mod provider_isolation;
-mod runtime_mismatch;
 mod tui_followup;
 mod tui_followup_retry;
 #[cfg(test)]
@@ -64,7 +63,6 @@ mod watchdog;
 
 use self::goal_lifecycle::*;
 use self::provider_isolation::*;
-use self::runtime_mismatch::*;
 use self::tui_followup::*;
 use self::tui_followup_retry::*;
 pub(in crate::services::discord) use self::turn_lifecycle::mailbox_try_start_turn_with_terminal_marker_cleanup;
